refactor(@formatjs/editor): rename component stories to fixtures
