+++ MODULE.bazel
@@ -431,6 +431,6 @@ http_file(
 bazel_dep(name = "rules_web_e2e", version = "0.0.0", dev_dependency = True)
 git_override(
     module_name = "rules_web_e2e",
-    commit = "b4efbe7448eace7fc0c936339198e152d7f03ef5",
+    commit = "aaa5079354f4b3510e8703235b2e26080e953ce1",
     remote = "https://github.com/perplexityai/rules_web_e2e.git",
 )
